Replace use of QwtText in metatype due to Qt unload bug
This change supports the new plugin reloading feature for pothos gui...

Qt does not support plugin unloading. This means that all 4 registered Qwt types
will have bad addresses when the modules are unloaded and reloaded because
Qt does not or cannot clear the old registration information.

./src/qwt_plot_item.h:Q_DECLARE_METATYPE( QwtPlotItem * )
./src/qwt_interval.h:Q_DECLARE_METATYPE( QwtInterval )
./src/qwt_graphic.h:Q_DECLARE_METATYPE( QwtGraphic )
./src/qwt_text.h:Q_DECLARE_METATYPE( QwtText )

We cant use these types as part of a metatype call.
Therefore all uses of QwtText were replaced with QString
and the styling was handled in the base class.
This change actually cleans up the code slightly.